# process-manager/src/main/webapp/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: Process Platform Manager API
  description: |
    API for applications using Process Platform for running its processes; part of the API is dedicated to Process Platform worker nodes

  version: "1.0"
servers:
  - url: /process-manager/api
...
paths:
...
  /profile/{profileId}:
    get:
      tags:
        - Profiles
      summary: Returns the specified profile
      description: Returns the specified profile
      parameters:
        - name: profileId
          in: path
          required: true
          description: Profile id
          schema:
            type: string
          example: testPlugin1-big
      responses:
        '200':
          description: JSON representation of the profile
          content:
            application/json:
              schema:
                type: string
                example: {
                  "profileId": "testPlugin1-big",
                  "description": null,
                  "pluginId": "testPlugin1",
                  "jvmArgs": [
                    "-Xms1g",
                    "-Xmx32g"
                  ]
                }
        '404':
          description: Profile does not exist
    put:
      tags:
        - Profiles
      summary: Updates the profile
      description: Updates the profile (just JVM args updated)
      parameters:
        - in: path
          name: profileId
          description: Profile id
          schema:
            type: string
          required: true
      requestBody:
        description: JSON representation of the profile
        required: true
        content:
          application/json:
            schema:
              type: string
            example: {
              "profileId": "testPlugin1-big",
              "description": null,
              "pluginId": "testPlugin1",
              "jvmArgs": [
                "-Xms1g",
                "-Xmx32g"
              ]
            }
      responses:
        '200':
          description: Successfully updated profile
        '400':
          description: Bad Request
  /process:
    post:
      tags:
        - Processes
      summary: Schedules new main process
      description: Schedules new main process
      requestBody:
        description: JSON with process information.
        required: true
        content:
          application/json:
            schema:
              type: string
            example: {
              "profileId" : "testPlugin1-big",
              "payload" : {
                "surname" : "Po",
                "name" : "Pe"
              },
              "ownerId" : "PePo"
            }
      responses:
        '200':
          description: New main process scheduled
          content:
            application/json:
              schema:
                type: string
              example: {"processId": "ed25ce29-2149-439d-85c4-cc5e516e3036"}
        '400':
          description: Bad Request
  /process/{processId}:
    get:
      tags:
        - Processes
      summary: Returns the process specified
      description: Returns the process specified
      parameters:
        - name: processId
          in: path
          required: true
          description: Process id
          schema:
            type: string
          example: ed25ce29-2149-439d-85c4-cc5e516e3036
      responses:
        '200':
          description: JSON representation of the process
          content:
            application/json:
              schema:
                type: string
                example: {
                  "processId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                  "description": "Main process for the profile testPlugin1-big",
                  "profileId": "testPlugin1-big",
                  "workerId": "curatorWorker",
                  "pid": 9889,
                  "planned": 1756198668715,
                  "started": 1756202186751,
                  "finished": null,
                  "status": "RUNNING",
                  "payload": {
                    "name": "Pe",
                    "surname": "Po"
                  },
                  "batchId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                  "owner": "PePo"
                }
        '404':
          description: Process not found
  /process/batch:
    get:
      tags:
        - Processes
      summary: Returns batches (groups od processes)
      description: Returns batches (groups of processes) allowing filtering and pagination
      parameters:
        - name: offset
          in: query
          required: false
          description: From where to start
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          description: How many rows to give back
          schema:
            type: integer
            default: 50
        - name: owner
          in: query
          required: false
          description: Filter by owner
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Filter by date from. (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Filter by date to. (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: state
          in: query
          required: false
          description: Filter by state
          schema:
            type: string
      responses:
        '200':
          description: List of batches
          content:
            application/json:
              schema:
                type: string
              example: {
                "batches": [
                  {
                    "owner": "PePo",
                    "processes": [
                      {
                        "owner": "PePo",
                        "workerId": "curatorWorker",
                        "processId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                        "payload": {
                          "surname": "Po",
                          "name": "Pe"
                        },
                        "profileId": "testPlugin1-big",
                        "description": "Main process for the profile testPlugin1-big",
                        "pid": 9889,
                        "started": 1756202186751,
                        "finished": null,
                        "planned": 1756198668715,
                        "batchId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                        "status": "RUNNING"
                      }
                    ],
                    "mainProcessId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                    "started": 1756202186751,
                    "finished": null,
                    "planned": 1756198668715,
                    "status": "RUNNING"
                  },
                  {
                    "owner": "PePo",
                    "processes": [
                      {
                        "owner": "PePo",
                        "workerId": "curatorWorker",
                        "processId": "6853579d-15c1-4fb9-ad11-3e107141aedb",
                        "payload": {
                          "surname": "Po",
                          "name": "Pe"
                        },
                        "profileId": "testPlugin1-big",
                        "description": "NewProcessName-PePo",
                        "pid": 27204,
                        "started": 1755262856918,
                        "finished": 1755262857087,
                        "planned": 1755262848459,
                        "batchId": "6853579d-15c1-4fb9-ad11-3e107141aedb",
                        "status": "FINISHED"
                      },
                      {
                        "owner": "PePo",
                        "workerId": "curatorWorker",
                        "processId": "114a8406-1788-4cc5-bb04-89ca1a8ba9fe",
                        "payload": {},
                        "profileId": "testPlugin2",
                        "description": "Sub process for the profile testPlugin2",
                        "pid": 10900,
                        "started": 1755262858373,
                        "finished": 1755262858428,
                        "planned": 1755262857057,
                        "batchId": "6853579d-15c1-4fb9-ad11-3e107141aedb",
                        "status": "FINISHED"
                      }
                    ],
                    "mainProcessId": "6853579d-15c1-4fb9-ad11-3e107141aedb",
                    "started": 1755262856918,
                    "finished": 1755262857087,
                    "planned": 1755262848459,
                    "status": "FINISHED"
                  }
                ],
                "totalSize": 2,
                "offset": 0,
                "limit": 50
              }
        '400':
          description: Bad request - wrong parameters
  /process/batch/{mainProcessId}:
    delete:
      tags:
        - Processes
      summary: Deletes a batch for given main process id
      description: Deletes a batch for a given process id. All processes in the batch must "deletable" - e.g. not running
      parameters:
        - name: mainProcessId
          in: path
          required: true
          description: Main process id
          example: ed25ce29-2149-439d-85c4-cc5e516e3036
      responses:
        '200':
          description: Batch deleted
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request - invalid state
  /process/batch/{mainProcessId}/execution:
    delete:
      tags:
        - Processes
      summary: Kills running batch processes
      description: Kills running batch processes
      parameters:
        - name: mainProcessId
          in: path
          required: true
          description: Main process Id
          example: 576
      responses:
        '200':
          description: Processes killed
        '404':
          description: Bad Request - invalid state
...
  /process/{processId}/log/out:
    get:
      tags:
        - Processes
      summary: Returns process standard output logs
      description: Returns process standard output logs (delegates to a worker)
      parameters:
        - name: processId
          in: path
          required: true
          description: Process id
          schema:
            type: string
            format: uuid
          example: 103cb713-b778-4370-8d58-f13bf4b276ec
        - name: fileName
          in: query
          required: false
          description: File name. Default "out.txt".
          schema:
            type: string
            default: "out.txt"
      responses:
        '200':
          description: Log provided
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Process logs not found for the given processId.
  /process/{processId}/log/err:
    get:
      tags:
        - Processes
      summary: Returns process error output logs
      description: Returns process error output logs (delegates to a worker)
      parameters:
        - name: processId
          in: path
          required: true
          description: uuid procesu
          schema:
            type: string
            format: uuid
          example: 103cb713-b778-4370-8d58-f13bf4b276ec
        - name: fileName
          in: query
          required: false
          description: File name. Default "err.txt".
          schema:
            type: string
            default: "err.txt"
      responses:
        '200':
          description: Log provided
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Error logs not found for the given process id.